Fix architecture detection on Windows in Makefile
$PROCESSOR_ARCHITECTURE seems to contain the architecture of the host,
but we need the architecture the current MinGW shell is targeting.
$MINGW_CHOST seems to be just that, and on my system it's either
i686-w64-mingw32 (mingw32.exe) or x86_64-w64-mingw32 (mingw64.exe)
(No idea what it looks like for Windows on ARM...)

Also enforce usage of SSE math on i386 and rename OSTYPE to YQ2_OSTYPE
and ARCH to YQ2_ARCH in the Makefile for a little more consistence
with the other Makefiles.